meer testen toegevoegd

+++ src/main/java/nl/nanda/transaction/dao/TransactionRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 
-    public Transaction findByEntityId(Integer id);
+    public Transaction findByEntityId(Long id);
 
-    public Transaction findByAccount(Integer id);
+    public Transaction findByAccount(Long id);
 
-    public Transaction findByTransfer(Integer id);
+    public Transaction findByTransfer(Long id);
 
     @Override
     public <S extends Transaction> S saveAndFlush(S arg0);
